Several "possibly controversial" corrections.
These are changes to make the HTML match the original, but the
original seems to have typographical errors. I have marked the
errors with a "<!-- sic -->" comment to indicate that it appears
that way in the original. I have also added some "sics" to
errors that were already reproduced in the HTML.